Add start and finish character handshaking Corrected custom filter setting from the script side
HWCDC does not work that well. Some bugs around Serial.end()/Serial.begin()
need to be resoved in Arduino ESP32 before it can be used. The lack of
DTR signalling makes things complicated. It is manageable with the addition
of some signalling characters in the reverse direction. It is now working
with private changes to Arduino ESP32. UCBCDC is still the prefered option.